pyproject.toml
[project]
name = "repoinsight-auth"
version = "0.1.0"
description = "RepoInsight auth service for local FastAPI and Cloudflare Workers deployment"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.111.0",
  "pydantic==2.7.1",
  "pydantic-settings==2.2.1",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.1.3",
  "python-multipart==0.0.9",
  "httpx==0.27.0",
  "libsql==0.1.11",
  "python-dotenv==1.0.1",
  "email-validator==2.1.1",
  "jinja2==3.1.4",
  "itsdangerous==2.2.0",
  "slowapi==0.1.9",
]

[tool.pywrangler]
entrypoint = "worker.py"

[dependency-groups]
dev = [
  "workers-py>=1.72.0",
  "workers-runtime-sdk",
]
